backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework-simplejwt
pyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv